refactor(constants): migrated from macro constants to typed constexpr

+++ include/clock.h
@@ -6,5 +6,5 @@ class Clock {
 public:
 	Clock();
 	void Reset();
-	double GetDTSec();
+	float GetDTSec();
 };
